Fix the tests, redo some schema slightly
