+++ packages/core/src/sheet.ts
@@ -24,6 +24,8 @@ import {
     SupportedLevel
 } from "@xivgear/xivmath/xivconstants";
 import {
+    ComputedSetStats,
+    ComputedSetStatsExport,
     DisplayGearSlotKey,
     EquippedItem,
     EquipSlotKey,
@@ -62,7 +64,6 @@ import {getDefaultSims, getRegisteredSimSpecs, getSimSpecByStub} from "./sims/si
 import {DUMMY_SHEET_MGR, SheetManager} from "./persistence/saved_sheets";
 import {CustomItem} from "./customgear/custom_item";
 import {CustomFood} from "./customgear/custom_food";
-import {statsSerializationProxy} from "@xivgear/xivmath/xivstats";
 import {isMateriaAllowed, materiaShortLabel} from "./materia/materia_utils";
 import {inflateSetExport} from "./util/sheet_utils";
 import {SpecialStatType} from "@xivgear/data-api-client/dataapi";
@@ -1795,3 +1796,75 @@ export const ExportTypes = {
     InternalSaveAs,
     FullStatsExport,
 } as const;
+
+/**
+ * Transform a ComputedSetStats into a form that serializes properly. That is, it serializes the getters rather
+ * than only the backing data. This is realistically what you would want out of the fulldata API endpoint.
+ *
+ * @param stats
+ */
+export function statsSerializationProxy(stats: ComputedSetStats): ComputedSetStatsExport {
+    // The purpose of this is that the fullstats API won't correctly serialize the ComputedSetStatsImpl normally.
+    // We care about the
+    return new Proxy(stats, {
+        get(target, prop, receiver) {
+            // Remove this field, it doesn't serialize well anyway
+            if (prop === 'jobStats') {
+                return undefined;
+            }
+            // Check if the property is a getter on the prototype chain
+            let descriptor = Object.getOwnPropertyDescriptor(target, prop as string);
+            let proto = Object.getPrototypeOf(target);
+
+            while (!descriptor && proto) {
+                descriptor = Object.getOwnPropertyDescriptor(proto, prop as string);
+                proto = Object.getPrototypeOf(proto);
+            }
+
+            if (descriptor && typeof descriptor.get === 'function') {
+                return descriptor.get.call(target);
+            }
+
+            return Reflect.get(target, prop, receiver);
+        },
+        ownKeys(target) {
+            const keys = new Set<string | symbol>();
+
+            let obj: object = target;
+            while (obj) {
+                Reflect.ownKeys(obj).forEach((key) => {
+                    if (key === 'jobStats') {
+                        return;
+                    }
+                    if (typeof key === 'string' && !key.startsWith('_')) {
+                        const descriptor = Object.getOwnPropertyDescriptor(obj, key);
+                        if (descriptor && typeof descriptor.get === 'function') {
+                            keys.add(key);
+                        }
+                    }
+                });
+                obj = Object.getPrototypeOf(obj);
+            }
+
+            return Array.from(keys);
+        },
+        getOwnPropertyDescriptor(target, prop) {
+            if (prop === 'jobStats') {
+                return undefined;
+            }
+            const descriptor = Object.getOwnPropertyDescriptor(target, prop)
+                || Object.getOwnPropertyDescriptor(Object.getPrototypeOf(target), prop);
+
+            if (descriptor
+                && typeof descriptor.get === 'function'
+                && typeof prop === 'string'
+                && !prop.startsWith('_')) {
+                return {
+                    enumerable: true,
+                    configurable: true,
+                };
+            }
+            return undefined;
+        },
+    }) as unknown as ComputedSetStatsExport;
+}
