+++ packages/xivmath/src/xivstats.ts
@@ -109,70 +109,6 @@ function clamp(min: number, max: number, value: number) {
 }
 
 /**
- * Transform a ComputedSetStats into a form that serializes properly. That is, it serializes the getters rather
- * than only the backing data. This is realistically what you would want out of the fulldata API endpoint.
- *
- * @param stats
- */
-export function statsSerializationProxy(stats: ComputedSetStats): ComputedSetStats {
-    // The purpose of this is that the fullstats API won't correctly serialize the ComputedSetStatsImpl normally.
-    // We care about the
-    return new Proxy(stats, {
-        get(target, prop, receiver) {
-            // Check if the property is a getter on the prototype chain
-            let descriptor = Object.getOwnPropertyDescriptor(target, prop as string);
-            let proto = Object.getPrototypeOf(target);
-
-            while (!descriptor && proto) {
-                descriptor = Object.getOwnPropertyDescriptor(proto, prop as string);
-                proto = Object.getPrototypeOf(proto);
-            }
-
-            if (descriptor && typeof descriptor.get === 'function') {
-                return descriptor.get.call(target);
-            }
-
-            return Reflect.get(target, prop, receiver);
-        },
-        ownKeys(target) {
-            const keys = new Set<string | symbol>();
-
-            let obj: object = target;
-            while (obj) {
-                Reflect.ownKeys(obj).forEach((key) => {
-                    if (typeof key === 'string' && !key.startsWith('_')) {
-                        const descriptor = Object.getOwnPropertyDescriptor(obj, key);
-                        if (descriptor && typeof descriptor.get === 'function') {
-                            keys.add(key);
-                        }
-                    }
-                });
-                obj = Object.getPrototypeOf(obj);
-            }
-
-            return Array.from(keys);
-        },
-        getOwnPropertyDescriptor(target, prop) {
-            const descriptor = Object.getOwnPropertyDescriptor(target, prop)
-                || Object.getOwnPropertyDescriptor(Object.getPrototypeOf(target), prop);
-
-            if (
-                descriptor
-                && typeof descriptor.get === 'function'
-                && typeof prop === 'string'
-                && !prop.startsWith('_')
-            ) {
-                return {
-                    enumerable: true,
-                    configurable: true,
-                };
-            }
-            return undefined;
-        },
-    });
-}
-
-/**
  * ComputedSetStats implementation.
  *
  * Unlike the old ComputedSetStats, this should not be modified. Rather, if a modified version is required,
